[Shruti]Add:Generics

+++ GenericMax/Maximum.cs
@@ -8,7 +8,7 @@ namespace GenericMax
 {
     public class Maximum
     {
-        public int Number(int firstvalue, int secondvalue, int thirdvalue)
+        public double Number(double firstvalue, double secondvalue, double thirdvalue)
         {
             //TestCase-1
             if (firstvalue.CompareTo(secondvalue) > 0 && firstvalue.CompareTo(thirdvalue) > 0 ||
@@ -33,8 +33,5 @@ namespace GenericMax
             }
             return firstvalue;
         }
-
-
     }
 }
-
